Extract common parts between mill versions
