+++ modules/mill-codegen-plugin/src-mill-shared/smithy4s/codegen/LSP.scala
@@ -25,12 +25,14 @@ import mmill.define.Target
 import mmill.eval.Evaluator
 import smithy4s.codegen.SmithyBuildJson
 import smithy4s.codegen.mill.Smithy4sModule
+import scala.annotation.nowarn
 
 import scala.collection.immutable.ListSet
 
 object LSP extends ExternalModule {
   lazy val millDiscover = mmill.define.Discover[this.type]
 
+  @nowarn("cat=deprecation")
   def updateConfig(ev: Evaluator): Command[PathRef] = {
     val rootPath = ev.rootModule.millModuleBasePath.value
     val s4sModules = ev.rootModule.millInternal.modules
